+++ board/milkv/vega/mem.go
@@ -16,5 +16,5 @@ import (
 
 // Applications can override ramSize with the `linkramsize` build tag.
 
-//go:linkname ramSize runtime/goos.RamSize
-var ramSize uint64 = 0x0f000000 // 240 MB
+//go:linkname ramSize github.com/usbarmory/tamago/goos.RamSize
+var ramSize uintptr = 0x0f000000 // 240 MB
