+++ board/nuclei/eval_soc/mem.go
@@ -20,5 +20,5 @@ import (
 //
 // Applications can override ramSize with the `linkramsize` build tag.
 
-//go:linkname ramSize runtime/goos.RamSize
-var ramSize uint64 = 0x0c800000 // 200 MB
+//go:linkname ramSize github.com/usbarmory/tamago/goos.RamSize
+var ramSize uintptr = 0x0c800000 // 200 MB
