+++ board/nxp/mx6ullevk/mem.go
@@ -21,5 +21,5 @@ import (
 
 // The MCIMX6ULL-EVK features a single 512MB DDR3 RAM module.
 
-//go:linkname ramSize runtime/goos.RamSize
-var ramSize uint32 = 0x20000000 // 512MB
+//go:linkname ramSize github.com/usbarmory/tamago/goos.RamSize
+var ramSize uintptr = 0x20000000 // 512MB
