+++ board/qemu/virt/mem.go
@@ -16,5 +16,5 @@ import (
 
 // Applications can override ramSize with the `linkramsize` build tag.
 //
-//go:linkname ramSize runtime/goos.RamSize
-var ramSize uint64 = 0x10000000 // 256MB
+//go:linkname ramSize github.com/usbarmory/tamago/goos.RamSize
+var ramSize uintptr = 0x10000000 // 256MB
