+++ board/raspberrypi/pi1/mem.go
@@ -18,12 +18,12 @@ import (
 // - model A  (all)
 // - model A+ (before 10th August 2016)
 // - model B  (before 15th October 2012)
-// var ramSize uint32 = 0x10000000 - 0x04000000 // 256MB - 64MB GPU (VideoCore)
+// var ramSize uintptr = 0x10000000 - 0x04000000 // 256MB - 64MB GPU (VideoCore)
 
 // the following models have 512 MB RAM:
 // - model A+ (after 10th August 2016)
 // - model B  (after 15th October 2012)
 // - model B+ (all)
 
-//go:linkname ramSize runtime/goos.RamSize
-var ramSize uint32 = 0x20000000 - 0x04000000 // 512MB - 64MB GPU (VideoCore)
+//go:linkname ramSize github.com/usbarmory/tamago/goos.RamSize
+var ramSize uintptr = 0x20000000 - 0x04000000 // 512MB - 64MB GPU (VideoCore)
