+++ soc/bcm2835/mem.go
@@ -14,5 +14,5 @@ import (
 	_ "unsafe"
 )
 
-//go:linkname ramStart runtime/goos.RamStart
-var ramStart uint32 = 0x00100000
+//go:linkname ramStart github.com/usbarmory/tamago/goos.RamStart
+var ramStart uintptr = 0x00100000
