+++ soc/nxp/imx8mp/init.go
@@ -25,8 +25,8 @@ const (
 	IMX8MPQ = 0x8240
 )
 
-//go:linkname ramStackOffset runtime/goos.RamStackOffset
-var ramStackOffset uint64 = 0x100
+//go:linkname ramStackOffset github.com/usbarmory/tamago/goos.RamStackOffset
+var ramStackOffset uintptr = 0x100
 
 var (
 	// Processor family
