+++ soc/nxp/imx8mp/mem.go
@@ -14,5 +14,5 @@ import (
 	_ "unsafe"
 )
 
-//go:linkname ramStart runtime/goos.RamStart
-var ramStart uint64 = DDR_BASE
+//go:linkname ramStart github.com/usbarmory/tamago/goos.RamStart
+var ramStart uintptr = DDR_BASE
